test: remove legacy signing key arg tests
Remove or replace in-toto-sign and in-toto-record cli tests related to
the the deprecated --key argument, and runlib tests related to the
deprecated signing_key argument.

Similar tests, which use the replacement --signing-key (cli) or signer
(runlib) argument have been added previously.

A few missing test cases are ported.

The commit also switches to the in-memory test key store, instead of
loading keys from files.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>